use a OneBased_IntegerVector instead of a std::vector<int>
